Split cstoken and stoken to be different concepts
The stokens are really just integrity checks for items, and are really
just tied to what revision we expected to have first what we have. So we
will rename stoken to lastRev or something, and have them completely
separate.

A sort of a revert of 65c6a131ac7684e920779db1c3e745855b75b954