Collection/item: remove content on delete (can be disabled).
More often than not, we want the content of the item deleted when we delete
an item. This means users can easily free up the space of the content on
deletion. It can always be accessed later through an old revision if needed.

Either way, there's a parameter to control this behaviour.