Revision: make mac calculation unambiguous.
The mac calculation was potentially ambiguous because we were
concatenating the data together without any delimiters, so an attacker
could craft two different structures with the same mac.
This is unlikely in practice due to the structure of the data, but it's
something we wanted to fix since this was first written.