Session storage: change format and make it possible to encrypt it.
Encrypting it lets people store the sensitive session data in a secure
manner.
The key should be a 32 bytes randomly generated buffer (can be created
using `randomBytes(32)`) and stored in the device's keystore or whatever
secure alternative. Alternatively, it can be derived from a user
password, though that's not yet documented.