Revert "Make chunk UIDs a merge of the nonce and the mac."
The idea behind this was that chunks could otherwise clash. Though it's
so extremely unlikely, especially since chunks need to only be unique
per item and not globally.

The major downside of merging them was that non prefetched items were
much fatter because the chunk ids were much larger which could be an
issue with large items. Which is why we are reverting this.

This reverts commit f098dfa127a3eaf8168d2349cb1a74818069ee9b.