Revision: remove salt field.
The idea behind it was that if items are reverted to an old state (by
setting back the chunks to whatever value they were and setting back the
encrypted meta to exactly what it was, there could be a clash because
the UID would be the same. Though if that's ever the case, it's enough
to just re-encrypt the meta to recalc it.