Revision: use the meta's tag as the revision mac/uid.
It was split before because it was cleaner/easier not to have to re-encrypt the
meta every time we want to set the content/delete. Though in practice we
almost always want to change meta anyway when doing these operations,
and merging them together simplifies some things and saves a lot of
space. This space saving adds up when storing a lot of small items.
Especially when storing a full cache locally on-device.