Padding: pad data to add data fingerprinting resistance.
Files, especially large ones, can often be identified (even when
encrypted) due to their unique size. If a user has a file sized:
2,715,254,784 and we know they are using Ubuntu, it's quite likely they
are using Ubuntu 20.04, as this is its ISO size.

By using padding we can make file sizes less unique and thus less likely
to leak information.

We chose the "padme"[1] padding scheme, with a minimum pad size of 32 to
achieve this, though in the future we plan on making it configurable.

[1]: https://www.petsymposium.org/2019/files/papers/issue4/popets-2019-0056.pdf