Use msgpack as the new transfer encoding for server communications
JSON is generally much faster in JS because it's implemented by the
javascript engine rather than in javascript.

However, because we handle a lot of binary data msgpack is much more space
efficient than JSON and would be much faster to decode on the server
(and on non-JS clients).
In addition, msgpack has much better type support than json, which makes
the code much cleaner and easier to make correct.