Shuffle chunks and handle the case where content is repeated in chunks.
This is a continuation of the previous commit. We need a way to know if
two chunks are exactly the same (e.g. a lot of zeros or a recurring
pattern) even though it's extremely unlikely that we'll have such long
repetitions.

Since we already have indices of what goes where, we use them to shuffle
the chunks. Shuffling the chunks further reduces the likelihood of
fingerprinting content based on file-sizes.