Collection transaction: simplify and fix.
In the case of a new collection we don't need to do anything, as
there's no stoken, so it was just unnecessary.
In the case of a change we were slightly wrong because we were
blocking the change for any change in the collection rather than
just the collection item itself.