Always encode the chunk indices in the first chunk.
The issue is that before this change there were corner cases in which we
weren't able to know if this is a single de-duplicated block or just a
single block. Now, when we always msgpack it, we can be sure about it.
In addition to fixing this issue, this gives us more flexibility for
future changes, and the overhead is tiny for most cases (and hidden by
padding for very small inputs).

This is a continuation to 84225f746056a49f01eb47af7e110231e54a5201