Chunking: move the indices header to its own chunk.
There was an issue that the uid of the chunk was based on the content of
chunk not including the index data.

This could break in some cases, for example:
Assume we have an item with 2 chunks. One with UID A and one with UID B.
We uploaded the file and the order it was uploaded to was: A, B.
This means that the indices were encoded into B, though because the UID
only depended on the content, it was uploaded with UID B.

We changed A so its UID is now A', and again, coming to upload the file
the shuffle turned up with the opposite order: B, A'.
However, since B's UID is the same, it's not going to be uploaded again
and thus it will include the msg packed content which is wrong! So on
the next download of the item we will get a corruption.

We solved this by not trying to be smart-asses anymore and just encode
the indices in their own chunks. It's a better idea anyway because it
provides us with more flexibility and the overhead is not significant.