Collection/Item: Make cache load and save sync.
It was async for extra future-proofing, though really, it was just
making things unnecessarily painful with little to no gain...
This operation is very lightweight anyway (just parse a small msgpack
buffer and populate the class), and if we want to change it in the
future, we can just introduce a new function.

For example, if we want to make it async, we can just add an async
variant. If we want to make it cache into a JSON (will be faster, and
potentially more compact if you know you are going to save a string), we
can just make a variant for that.