Collection: merged CollectionMetadata with ItemMetadata.
We made them different because the allowed keys were slightly different,
and we wanted to enforce name and type to be set for collections.

However, we are going to move the collection type out of the collections
so enforcing it to be set is no longer needed, and while we still want
to encourage people to set the name, it's not really a requirement and
not worth having the two types.