Collection Type: change the tag to use deterministic encryption rather than mac.
Using deterministic encryption we always get the same ciphertext for the same
plaintext. This makes the encryption tag reversible (so we can get the
string value from the token) for a fairly small cost.

The mac method is a bit faster to compute (as it's just a mac) and the
tag size is 32 bytes.
With the deterministic encryption method, we have 24 bytes for the
nonce, 16 bytes for the authentication tag, and then the actual content
(padded to 32 bytes blocked). So in most cases, the tag size would be 72
bytes.

However, being able to get the type back makes the API much nicer and
stores useful information, so it's worth it.