Async: change all of the meta/delete/colType functions sync.
The idea behind async is that we could move them to a service worker
transparently. This idea wasn't going to work though as jumping back and
forth between the service worker on every such short call would have
been too slow.

Additionally, this was causing major headaches because you can only
`await` in async functions so running even short operations that we can
afford to wait on was a pain.